+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.3.3] - in progress
 
-- New module `cloudly.gcp.cloudsql` with support for Postgres, PGVector.
+- New module `cloudly.gcp.sql.postgres`.
+- New module `cloudly.sql.postgres`.
 
 
 ## [0.3.2] - 2025-01-28